Stop upward sand creep in deserts.
This is an extreme measure to stop upward sand creep in non-deserts.
If there isn't some sand already at the same elevation as the eroding
dirt, don't erode it to sand. This prevents beaches slowly crawling up
entire hills.

This is needed since there is no biome information yet in Minetest.